Add utility function to extract measurement metadata
